Add basic nav menu
